+++ oct2py/__init__.py
@@ -9,7 +9,7 @@ MAT files.  Usage is as simple as:
     >>> import oct2py
     >>> oc = oct2py.Oct2Py()
     >>> x = oc.zeros(3,3)
-    >>> print(x, x.dtype.str)
+    >>> print(x, x.dtype.str)  # doctest: +SKIP
     [[ 0.  0.  0.]
      [ 0.  0.  0.]
      [ 0.  0.  0.]] <f8
@@ -18,9 +18,8 @@ If you want to run legacy m-files, do not have MATLAB(TM), and do not fully
 trust a code translator, this is your library.
 """
 
-
 __title__ = 'oct2py'
-__version__ = '2.4.1'
+__version__ = '3.1.0'
 __author__ = 'Steven Silvester'
 __license__ = 'MIT'
 __copyright__ = 'Copyright 2014 Steven Silvester'
@@ -45,15 +44,18 @@ if os.name == 'nt':
     http://stackoverflow.com/questions/15457786/ctrl-c-crashes-python-after-importing-scipy-stats
     """
     basepath = imp.find_module('numpy')[1]
-    lib1 = ctypes.CDLL(os.path.join(basepath, 'core', 'libmmd.dll'))
-    lib2 = ctypes.CDLL(os.path.join(basepath, 'core', 'libifcoremd.dll'))
+    try:
+        lib1 = ctypes.CDLL(os.path.join(basepath, 'core', 'libmmd.dll'))
+        lib2 = ctypes.CDLL(os.path.join(basepath, 'core', 'libifcoremd.dll'))
 
-    def handler(sig, hook=thread.interrupt_main):
-        hook()
-        return 1
+        def handler(sig, hook=thread.interrupt_main):
+            hook()
+            return 1
 
-    routine = ctypes.WINFUNCTYPE(ctypes.c_int, ctypes.c_uint)(handler)
-    ctypes.windll.kernel32.SetConsoleCtrlHandler(routine, 1)
+        routine = ctypes.WINFUNCTYPE(ctypes.c_int, ctypes.c_uint)(handler)
+        ctypes.windll.kernel32.SetConsoleCtrlHandler(routine, 1)
+    except Exception:
+        pass
 
 
 from .core import Oct2Py, Oct2PyError
@@ -88,4 +90,3 @@ def kill_octave():
 
 # clean up namespace
 del functools, imp, os, ctypes, thread
-del core, utils