+++ oct2py/matwrite.py
@@ -11,7 +11,7 @@ import os
 from scipy.io import savemat
 import numpy as np
 from scipy.sparse import csr_matrix, csc_matrix
-from .utils import Oct2PyError, create_file
+from .utils import Oct2PyError
 from .compat import unicode
 
 
@@ -20,12 +20,11 @@ class MatWrite(object):
 
     Strives to preserve both value and type in transit.
     """
-    def __init__(self, temp_dir=None, oned_as='row'):
+    def __init__(self, oned_as='row', convert_to_float=True):
         self.oned_as = oned_as
-        self.temp_dir = temp_dir
-        self.in_file = create_file(self.temp_dir)
+        self.convert_to_float = convert_to_float
 
-    def create_file(self, inputs, names=None):
+    def create_file(self, temp_dir, inputs, names=None):
         """
         Create a MAT file, loading the input variables.
 
@@ -59,14 +58,13 @@ class MatWrite(object):
             # for structs - recursively add the elements
             try:
                 if isinstance(var, dict):
-                    data[argin_list[-1]] = putvals(var)
+                    data[argin_list[-1]] = putvals(var, self.convert_to_float)
                 else:
-                    data[argin_list[-1]] = putval(var)
+                    data[argin_list[-1]] = putval(var, self.convert_to_float)
             except Oct2PyError:
                 raise
             ascii_code += 1
-        if not os.path.exists(self.in_file):
-            self.in_file = create_file(self.temp_dir)
+        self.in_file = os.path.join(temp_dir, 'writer.mat')
         try:
             savemat(self.in_file, data, appendmat=False,
                     oned_as=self.oned_as, long_field_names=True)
@@ -76,14 +74,8 @@ class MatWrite(object):
                                           '" "'.join(argin_list))
         return argin_list, load_line
 
-    def remove_file(self):
-        try:
-            os.remove(self.in_file)
-        except (OSError, AttributeError):  # pragma: no cover
-            pass
 
-
-def putvals(dict_):
+def putvals(dict_, convert_to_float=True):
     """
     Put a nested dict into the MAT file as a struct
 
@@ -91,6 +83,8 @@ def putvals(dict_):
     ==========
     dict_ : dict
         Dictionary of object(s) to store
+    convert_to_float : bool
+        If true, convert integer types to float
 
     Returns
     =======
@@ -101,13 +95,13 @@ def putvals(dict_):
     data = dict()
     for key in dict_.keys():
         if isinstance(dict_[key], dict):
-            data[key] = putvals(dict_[key])
+            data[key] = putvals(dict_[key], convert_to_float)
         else:
-            data[key] = putval(dict_[key])
+            data[key] = putval(dict_[key], convert_to_float)
     return data
 
 
-def putval(data):
+def putval(data, convert_to_float=True):
     """
     Convert data into a state suitable for transfer.
 
@@ -115,6 +109,8 @@ def putval(data):
     ==========
     data : object
         Value to write to file.
+    convert_to_float : bool
+        If true, convert integer types to float.
 
     Returns
     =======
@@ -178,6 +174,8 @@ def putval(data):
         raise Oct2PyError('Datatype not supported: {0}'.format(data.dtype))
     if data.dtype == 'object' and len(data.shape) > 1:
         data = data.T
+    if convert_to_float and data.dtype.kind in 'uib':
+        data = data.astype(float)
     return data